// src/api/__generated__/wallet/model/dayCheckinVO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 单日签到信息
 */
export interface DayCheckinVO {
  /** 周几 */
  dayOfWeek?: number;
  /** 签到日期 */
  date?: string;
  /** 获得积分 */
  points?: number;
  /** 是否已签到 */
  isSigned?: boolean;
}