// src/api/__generated__/wallet/model/pointsLedgerItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { PointsLedgerItemSourceType } from './pointsLedgerItemSourceType';

/**
 * 积分流水单项
 */
export interface PointsLedgerItem {
  /** 积分来源类型：CHECKIN=每日签到，FOLLOW_X=关注Twitter，RETWEET=转发推文，LIKE=点赞推文，SHARE_X=Twitter传播加分，SHARE_TIKTOK=TikTok传播加分，SHARE_YOUTUBE=YouTube传播加分，PUBLISH_DRAMA=发布短剧，FINISH_DRAMA=完播短剧，DISPATCH_MINING=派遣演员挖矿，BIND_X=绑定Twitter，FINAL_SNAPSHOT=活动结束积分快照 */
  sourceType?: PointsLedgerItemSourceType;
  /** 本次积分变动(增量) */
  points?: number;
  /** 入账时间戳(ms) */
  createdAt?: number;
}