// src/api/__generated__/wallet/model/pointsLedgerItemSourceType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 积分来源类型：CHECKIN=每日签到，FOLLOW_X=关注Twitter，RETWEET=转发推文，LIKE=点赞推文，SHARE_X=Twitter传播加分，SHARE_TIKTOK=TikTok传播加分，SHARE_YOUTUBE=YouTube传播加分，PUBLISH_DRAMA=发布短剧，FINISH_DRAMA=完播短剧，DISPATCH_MINING=派遣演员挖矿，BIND_X=绑定Twitter，FINAL_SNAPSHOT=活动结束积分快照
 */
export type PointsLedgerItemSourceType =
  (typeof PointsLedgerItemSourceType)[keyof typeof PointsLedgerItemSourceType];

export const PointsLedgerItemSourceType = {
  CHECKIN: 'CHECKIN',
  FOLLOW_X: 'FOLLOW_X',
  RETWEET: 'RETWEET',
  LIKE: 'LIKE',
  SHARE_X: 'SHARE_X',
  SHARE_TIKTOK: 'SHARE_TIKTOK',
  SHARE_YOUTUBE: 'SHARE_YOUTUBE',
  PUBLISH_DRAMA: 'PUBLISH_DRAMA',
  FINISH_DRAMA: 'FINISH_DRAMA',
  DISPATCH_MINING: 'DISPATCH_MINING',
  BIND_X: 'BIND_X',
  FINAL_SNAPSHOT: 'FINAL_SNAPSHOT',
} as const;