// src/api/__generated__/wallet/model/shareAttemptResponseStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 审核状态
 */
export type ShareAttemptResponseStatus =
  (typeof ShareAttemptResponseStatus)[keyof typeof ShareAttemptResponseStatus];

export const ShareAttemptResponseStatus = {
  pending_approval: 'pending_approval',
  approved: 'approved',
  rejected: 'rejected',
} as const;