// src/api/__generated__/wallet/model/taskStatusItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { TaskStatusItemStatus } from './taskStatusItemStatus';

/**
 * 任务状态项
 */
export interface TaskStatusItem {
  /** 任务 ID（后管分配，10–99） */
  taskId?: number;
  /** 任务类型 */
  type?: string;
  /** 完成可得积分 */
  points?: number;
  /** 任务跳转链接 */
  linkUrl?: string;
  /** 任务描述 */
  description?: string;
  /** 任务状态 */
  status?: TaskStatusItemStatus;
}