// src/api/__generated__/wallet/model/taskStatusItemStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 任务状态
 */
export type TaskStatusItemStatus =
  (typeof TaskStatusItemStatus)[keyof typeof TaskStatusItemStatus];

export const TaskStatusItemStatus = {
  NOT_DONE: 'NOT_DONE',
  DONE: 'DONE',
  CLAIMED: 'CLAIMED',
} as const;